break result card tabs into components to lazy load result details
show loading state
